Change fTrigger ID type from int to unsigned long to record TimeStamp

+++ dataModel/PixelTPCdata.cpp
@@ -21,7 +21,7 @@ PixelTPCdata::PixelTPCdata(int NumofChips)
     }
 }
 
-PixelTPCdata& PixelTPCdata::SetTiggleID(int triggleid)
+PixelTPCdata& PixelTPCdata::SetTiggleID(unsigned long triggleid)
 {
     fTriggleID = triggleid;
     return *this;
